Moved `csv.js` to `utils` folder and rename to `ts`
