+++ src/popup/tabs/download_tab.js
@@ -1,6 +1,6 @@
 import { Release } from "../../app/release.js";
 import { DiscogsCsv } from "../../discogs/app/discogs-csv.js";
-import { downloadCsv, objectsToCsv } from "../../modules/csv.js";
+import { downloadCsv, objectsToCsv } from "../../utils/csv";
 import { isEmptyArray } from "../../modules/utils.js";
 import { disable, enable } from "../../modules/html.js";
